+++ src/main/java/com/hzkans/crm/modules/wechat/entity/WechatReplyKeywordDO.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class WechatReplyKeywordDO extends DataEntity<WechatReplyKeywordDO> {
 
 	private static final long serialVersionUID = 1L;
-	private Integer wechatId;		// 微信ID
+	private Long wechatId;		// 微信ID
 	private String ruleId;		// 规则ID
 	private Integer keywordType;		// 关键词类型： 0-半匹配  1-全匹配
 	private String keyword;		// 关键词
@@ -49,14 +49,13 @@ public class WechatReplyKeywordDO extends DataEntity<WechatReplyKeywordDO> {
 		}
 	}
 
-	public Integer getWechatId() {
+	public Long getWechatId() {
 		return wechatId;
 	}
 
-	public void setWechatId(Integer wechatId) {
+	public void setWechatId(Long wechatId) {
 		this.wechatId = wechatId;
 	}
-
 
 	public String getRuleId() {
 		return ruleId;
