+++ src/main/java/com/hzkans/crm/modules/wechat/service/WechatPlatfromService.java
@@ -11,13 +11,13 @@ import java.util.List;
  * Time:17:47
  */
 public interface WechatPlatfromService {
-    WechatPlatfromDO getWechatPlatformById(Integer id) throws Exception;
+    WechatPlatfromDO getWechatPlatformById(Long id) throws Exception;
 
     void addWechatPlatform(WechatPlatfromDO wechatPlatfromDO) throws Exception;
 
     void updateWechatPlatform(WechatPlatfromDO wechatPlatfromDO) throws Exception;
 
-    void removeWechatPlatform(Integer id) throws Exception;
+    void removeWechatPlatform(Long id) throws Exception;
 
     List<WechatPlatfromDO> listWechatPlatform(WechatPlatfromDO wechatPlatfromDO) throws Exception;
